Revert "extract create wiki dialog to component"
This reverts commit c7bff1b4ed52a42befd11c18bbaa147ca8a82114.
